Added the missing virtual destructors. Changed exception string formatting in DefaultNearCache so that there is no warning at 32-bit compiler.

+++ hazelcast/include/hazelcast/client/internal/nearcache/impl/DefaultNearCache.h
@@ -151,7 +151,7 @@ namespace hazelcast {
                                                                                             serializationService));
                                 default:
                                     std::ostringstream out;
-                                    out << "Invalid in memory format: " << inMemoryFormat;
+                                    out << "Invalid in memory format: " << (int) inMemoryFormat;
                                     throw exception::IllegalArgumentException(out.str());
                             }
                         }
